+++ src/compositor/mod.rs
@@ -13,7 +13,7 @@ pub trait CompositorAttr {
 
 /// Compositor composites two pixels with alpha value.
 pub trait Compositor<T>: CompositorAttr {
-    fn composite(&self, dst: &T, src: &T, alpha: f64) -> T;
+    fn composite(&self, dst: &T, src: &T, alpha: f32) -> T;
 }
 
 #[derive(Clone)]
@@ -94,30 +94,126 @@ pub enum Basic {
     Exclusion,
 }
 
-impl CompositorAttr for Clear {fn keep_dst_on_transparent_src(&self) -> bool {false}}
-impl CompositorAttr for Src {fn keep_dst_on_transparent_src(&self) -> bool {false}}
-impl CompositorAttr for Dst {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for SrcOver {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for SrcIn {fn keep_dst_on_transparent_src(&self) -> bool {false}}
-impl CompositorAttr for SrcOut {fn keep_dst_on_transparent_src(&self) -> bool {false}}
-impl CompositorAttr for SrcAtop {fn keep_dst_on_transparent_src(&self) -> bool {false}}
-impl CompositorAttr for DstOver {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for DstIn {fn keep_dst_on_transparent_src(&self) -> bool {false}}
-impl CompositorAttr for DstOut {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for DstAtop {fn keep_dst_on_transparent_src(&self) -> bool {false}}
-impl CompositorAttr for Xor {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Add {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Darken {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Lighten {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Multiply {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Screen {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Overlay {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for HardLight {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Dodge {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Burn {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for SoftLight {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Difference {fn keep_dst_on_transparent_src(&self) -> bool {true}}
-impl CompositorAttr for Exclusion {fn keep_dst_on_transparent_src(&self) -> bool {true}}
+impl CompositorAttr for Clear {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        false
+    }
+}
+impl CompositorAttr for Src {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        false
+    }
+}
+impl CompositorAttr for Dst {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for SrcOver {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for SrcIn {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        false
+    }
+}
+impl CompositorAttr for SrcOut {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        false
+    }
+}
+impl CompositorAttr for SrcAtop {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        false
+    }
+}
+impl CompositorAttr for DstOver {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for DstIn {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        false
+    }
+}
+impl CompositorAttr for DstOut {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for DstAtop {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        false
+    }
+}
+impl CompositorAttr for Xor {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Add {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Darken {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Lighten {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Multiply {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Screen {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Overlay {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for HardLight {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Dodge {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Burn {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for SoftLight {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Difference {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
+impl CompositorAttr for Exclusion {
+    fn keep_dst_on_transparent_src(&self) -> bool {
+        true
+    }
+}
 
 impl CompositorAttr for Basic {
     fn keep_dst_on_transparent_src(&self) -> bool {
@@ -153,8 +249,8 @@ impl CompositorAttr for Basic {
 
 /// For Measuring performance.
 pub mod perf {
-    use crate::pixel::Rgba;
     use super::*;
+    use crate::pixel::Rgba;
 
     #[derive(Clone)]
     pub struct Perf;
@@ -168,7 +264,12 @@ pub mod perf {
     #[cfg(feature = "image-crate")]
     impl Compositor<image::Rgba<u8>> for Perf {
         #[allow(unused_variables)]
-        fn composite(&self, a: &image::Rgba<u8>, b: &image::Rgba<u8>, alpha: f64) -> image::Rgba<u8> {
+        fn composite(
+            &self,
+            a: &image::Rgba<u8>,
+            b: &image::Rgba<u8>,
+            alpha: f32,
+        ) -> image::Rgba<u8> {
             image::Rgba([a.0[0], b.0[0], alpha as u8, 255])
         }
     }
@@ -176,14 +277,14 @@ pub mod perf {
     #[cfg(feature = "image-crate")]
     impl Compositor<image::Rgb<u8>> for Perf {
         #[allow(unused_variables)]
-        fn composite(&self, a: &image::Rgb<u8>, b: &image::Rgb<u8>, alpha: f64) -> image::Rgb<u8> {
+        fn composite(&self, a: &image::Rgb<u8>, b: &image::Rgb<u8>, alpha: f32) -> image::Rgb<u8> {
             image::Rgb([a.0[0], b.0[0], alpha as u8])
         }
     }
 
     impl Compositor<Rgba> for Perf {
         #[allow(unused_variables)]
-        fn composite(&self, a: &Rgba, b: &Rgba, alpha: f64) -> Rgba {
+        fn composite(&self, a: &Rgba, b: &Rgba, alpha: f32) -> Rgba {
             Rgba([a.0[0], b.0[0], alpha as f32, 1.0])
         }
     }
