+++ src/fill_color/mod.rs
@@ -17,7 +17,7 @@ pub use transform::Transform;
 /// A trait for Colorization.
 /// Generates `Pixel` from (x, y).
 pub trait FillColor<C> {
-    fn fill_color(&self, x: f64, y: f64) -> C;
+    fn fill_color(&self, x: f32, y: f32) -> C;
 }
 
 /// Solid color.
@@ -31,23 +31,23 @@ impl<C: Clone> Solid<C> {
 }
 
 impl<C: Clone> FillColor<C> for Solid<C> {
-    fn fill_color(&self, _: f64, _: f64) -> C {
+    fn fill_color(&self, _: f32, _: f32) -> C {
         self.0.clone()
     }
 }
 
 /// Block check pattern.
 #[derive(Debug, Clone)]
-pub struct BlockCheck<C: Clone>(C, C, f64);
+pub struct BlockCheck<C: Clone>(C, C, f32);
 
 impl<C: Clone> BlockCheck<C> {
-    pub fn new(c1: C, c2: C, size: f64) -> Self {
+    pub fn new(c1: C, c2: C, size: f32) -> Self {
         BlockCheck(c1, c2, size)
     }
 }
 
 impl<C: Clone> FillColor<C> for BlockCheck<C> {
-    fn fill_color(&self, x: f64, y: f64) -> C {
+    fn fill_color(&self, x: f32, y: f32) -> C {
         if ((x / self.2) as isize + (y / self.2) as isize) % 2 == 0 {
             self.0.clone()
         } else {
